added makedir

+++ scripts/core_build.py
@@ -654,22 +654,24 @@ def write_results_to_file(changes_dict):
     print(f"All the data for build has been written to {testPath}/built_changes.json")
 
 def main():
-    # Generate clocks.json
-    input_directory = "./"
-    output_dir = "./output/docs"
-    output_file = os.path.join(output_dir, 'clocks.json')
-    clocksGenerator = GenerateClocks(input_directory, output_file)
-    clocksGenerator.generate()
+    os.makedirs(testPath, exists_ok = True)
+    if os.name == 'linux':
+        # Generate clocks.json
+        input_directory = "./"
+        output_dir = "./output/docs"
+        output_file = os.path.join(output_dir, 'clocks.json')
+        clocksGenerator = GenerateClocks(input_directory, output_file)
+        clocksGenerator.generate()
 
-    # Remove the existing clocks.json file if it exists
-    clocks_path = os.path.join(local_app_data_path, 'clocks.json')
-    if os.path.exists(clocks_path):
-        os.remove(clocks_path)
-        shutil.copy(output_file, local_app_data_path)
-        shutil.copy(output_file, testPath)
-        print(f"\033[93mReplaced {clocks_path} with: {output_file}\033[0m")
-    else:
-        print(f"\033[91mFile not found: {clocks_path}\033[0m")
+        # Remove the existing clocks.json file if it exists
+        clocks_path = os.path.join(local_app_data_path, 'clocks.json')
+        if os.path.exists(clocks_path):
+            os.remove(clocks_path)
+            shutil.copy(output_file, local_app_data_path)
+            shutil.copy(output_file, testPath)
+            print(f"\033[93mReplaced {clocks_path} with: {output_file}\033[0m")
+        else:
+            print(f"\033[91mFile not found: {clocks_path}\033[0m")
 
     files = get_changed_files('main')
     archs = []
